test: Test for a11y scale should not break style properties tests
